Make observed genes unique
Two intervals might be annotated with the same gene, which means that
the gene lists returned by annotate may contain duplicates. These were
included in the "observed" count per GO category. My feeling is, this is
a recent bug and that older versions of ego used maps to store gene IDs
rather than append to a slice, as is done in the last version. I've now
reverted to a map of IDs.